refactoring
break big component
